Started using Index Buffers, 24 vertices instead of 36.

+++ OpenGL1/src/camera.h
@@ -6,11 +6,11 @@ glm::vec3 camUp = glm::vec3(0.0f, 0.1f, 0.0f);
 float dt = 0.0f;
 float lastFrame = 0.0f;
 
-double lastX = 400 , lastY = 300;
-double xoff = 0.0f, yoff = 0.0f;
+float lastX = 400 , lastY = 300;
+float xoff = 0.0f, yoff = 0.0f;
 
 bool firstMouse = true;
 
-double mouse_sensitivity = 0.01f;
+float mouse_sensitivity = 0.01f;
 
-double yaw = 0.0f, pitch = -89.0f;
+float yaw = 0.0f, pitch = -89.0f;
